Restructure dependency handling.
The file containing the specializations of traits for specific optional
dependencies have been moved to their own files. This allows us to move
to macro free dependency handling, simply by removing certain guarded
includes.
