+++ src/examples/create_dataset_half_float.cpp
@@ -7,8 +7,6 @@
  *
  */
 
-#ifdef H5_USE_HALF_FLOAT
-
 #include <iostream>
 #include <string>
 #include <vector>
@@ -45,5 +43,3 @@ int main(void) {
 
     return 0;
 }
-
-#endif
