+++ tests/unit/tests_high_five.hpp
@@ -43,8 +43,6 @@ using base_test_types = std::tuple<int,
                                    fcomplex>;
 
 #ifdef H5_USE_HALF_FLOAT
-#include <half.hpp>
-
 using float16_t = half_float::half;
 using numerical_test_types =
     decltype(std::tuple_cat(std::declval<base_test_types>(), std::tuple<float16_t>()));
